wished load on user login

+++ pages/api/wishlist.js
@@ -5,26 +5,36 @@ import { authOptions } from "./auth/[...nextauth]";
 
 export default async function handle(req, res) {
   await mongooseConnect();
-  const { user } = await getServerSession(req, res, authOptions);
+  const session = await getServerSession(req, res, authOptions);
+
   if (req.method === "POST") {
-    const { product } = req.body;
-    const wishedDoc = await WishedProduct.findOne({
-      userEmail: user.email,
-      product
-    });
-    if (wishedDoc) {
-      await WishedProduct.findByIdAndDelete(wishedDoc._id);
-      res.json("wish deleted");
+    if (session && session.user) {
+      const { product } = req.body;
+      const wishedDoc = await WishedProduct.findOne({
+        userEmail: session.user.email,
+        product
+      });
+      if (wishedDoc) {
+        await WishedProduct.findByIdAndDelete(wishedDoc._id);
+        res.json("wish deleted");
+      } else {
+        await WishedProduct.create({ userEmail: session.user.email, product });
+        res.json("wish created");
+      }
     } else {
-      await WishedProduct.create({ userEmail: user.email, product });
-      res.json("wish created");
+      res.json(null);
     }
   }
+
   if (req.method == "GET") {
-    res.json(
-      await WishedProduct.find({ userEmail: user.email })
-        .populate("product")
-        .exec()
-    );
+    if (session && session.user) {
+      res.json(
+        await WishedProduct.find({ userEmail: session.user.email })
+          .populate("product")
+          .exec()
+      );
+    } else {
+      res.json(null);
+    }
   }
 }
